refactor: errors are defined with thiserror
