+++ src/domain/repository/sound_repository.rs
@@ -1,6 +1,6 @@
 
 use crate::domain::entity::sound::Sound;
-use crate::domain::error::kind::ProcessError;
+use crate::domain::error::ProcessError;
 
 pub trait SoundRepository {
     fn find_all(&self) -> Result<Vec<Sound>, ProcessError>;
