+++ src/domain/service/job_creation_service.rs
@@ -1,6 +1,7 @@
 use chrono::Utc;
 
-use crate::domain::error::kind::ProcessError;
+//use crate::domain::error::kind::ProcessError;
+use crate::domain::error::ProcessError;
 use std::io::BufReader;
 use tokio_cron_scheduler::Job;
 use crate::PlaySound;
@@ -8,13 +9,13 @@ use crate::PlaySound;
 pub struct JobCreateService {}
 
 fn play_sound(file: &std::fs::File) -> Result<(), ProcessError> {
-    let file = file.try_clone().map_err(|_| ProcessError::FileReadError)?;
+    let file = file.try_clone().map_err(|e| ProcessError::FileReadError(e))?;
     let decoder =
-        rodio::Decoder::new(BufReader::new(file)).map_err(|_| ProcessError::FileReadError)?;
+        rodio::Decoder::new(BufReader::new(file)).map_err(|e| ProcessError::DecodeError(e))?;
     // OutputStream (_stream) is needed to play sound.
     let (_stream, handle) =
-        rodio::OutputStream::try_default().map_err(|_| ProcessError::OutputStreamError)?;
-    let sink = rodio::Sink::try_new(&handle).map_err(|_| ProcessError::OutputStreamError)?;
+        rodio::OutputStream::try_default().map_err(|e| ProcessError::OutputStreamError(e))?;
+    let sink = rodio::Sink::try_new(&handle).map_err(|e| ProcessError::PlayingError(e))?;
 
     sink.append(decoder);
     sink.sleep_until_end();
