+++ src/use_case/sound.rs
@@ -1,6 +1,6 @@
 
 use crate::domain::entity::sound::Sound;
-use crate::domain::error::kind::ProcessError;
+use crate::domain::error::ProcessError;
 
 pub trait SoundUseCase {
     fn find_all_sounds(&self) -> Result<Vec<Sound>, ProcessError>;
